Prefer PlatformLocalStorage Over Other Native Modules
We currently try to load builtin native modules before trying to
fallback to platform provided native modules. Reversing this order
allows the platform to provide the AsyncStorage implementtion even if
on a platform that async-storage already supports. This ability can
be useful in hosted scenarios where we may want more control over how
data for an individual React Native instance is stored.