ci(iOS): Fix yarn install failing with Xcode 12 (#460)
Note that this does not fix tests failing to run on iOS 14. EarlGrey,
which is what Detox depends on, crashes when run on iOS 14
(https://github.com/google/EarlGrey/issues/1447). We may be required to
bump to latest Detox to solve this issue, but it will require more work
on our side.